DODANE NOTATKI I EDYCJA UZYTKOWNIKA

+++ KluczToSukcesDoKariery/Controllers/NotatkisController.cs
@@ -1,15 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using KluczToSukcesDoKariery.Data;
 using KluczToSukcesDoKariery.Models;
 
 namespace KluczToSukcesDoKariery.Controllers
 {
+    [Authorize]
     public class NotatkisController : Controller
     {
         private readonly KluczToSukcesDoKarieryContext _context;
@@ -22,9 +22,11 @@ namespace KluczToSukcesDoKariery.Controllers
         // GET: Notatkis
         public async Task<IActionResult> Index()
         {
-              return _context.Notatki != null ? 
-                          View(await _context.Notatki.ToListAsync()) :
-                          Problem("Entity set 'KluczToSukcesDoKarieryContext.Notatki'  is null.");
+            var userId = User.Identity.Name; // Pobierz ID zalogowanego użytkownika
+            var userNotes = await _context.Notatki
+                .Where(n => n.UserId == userId) // Filtrowanie notatek na podstawie UserId
+                .ToListAsync();
+            return View(userNotes);
         }
 
         // GET: Notatkis/Details/5
@@ -35,9 +37,10 @@ namespace KluczToSukcesDoKariery.Controllers
                 return NotFound();
             }
 
-            var notatki = await _context.Notatki
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (notatki == null)
+            var notatki = await _context.Notatki.FirstOrDefaultAsync(m => m.Id == id);
+
+            // Sprawdzenie, czy użytkownik jest właścicielem notatki
+            if (notatki == null || notatki.UserId != User.Identity.Name)
             {
                 return NotFound();
             }
@@ -46,20 +49,42 @@ namespace KluczToSukcesDoKariery.Controllers
         }
 
         // GET: Notatkis/Create
+        [HttpGet]
         public IActionResult Create()
         {
-            return View();
+            var notatki = new Notatki
+            {
+                UserId = User.Identity.Name, // Automatyczne przypisanie UserId
+                DataDodania = DateTime.Now
+                
+            };
+
+            return View(notatki);
         }
 
+
         // POST: Notatkis/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Tytul,Tresc,DataDodania,UserId")] Notatki notatki)
+        public async Task<IActionResult> Create([Bind("Tytul,Tresc,DataDodania,UserId")] Notatki notatki)
         {
+            // Sprawdzenie, czy użytkownik już posiada notatkę o tym samym tytule
+            var userId = User.Identity.Name;
+            bool noteExists = await _context.Notatki
+                .AnyAsync(n => n.Tytul == notatki.Tytul && n.UserId == userId);
+
+            if (noteExists)
+            {
+                ModelState.AddModelError("Tytul", "Notatka o takim tytule już istnieje.");
+                notatki.DataDodania = DateTime.Now;
+                return View(notatki); // Zwróć widok, jeśli tytuł już istnieje
+            }
+
+
             if (ModelState.IsValid)
             {
+                notatki.UserId = userId; // Przypisanie UserId
+                notatki.DataDodania = DateTime.Now; // Ustawienie aktualnej daty
                 _context.Add(notatki);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -76,21 +101,22 @@ namespace KluczToSukcesDoKariery.Controllers
             }
 
             var notatki = await _context.Notatki.FindAsync(id);
-            if (notatki == null)
+
+            // Sprawdzenie, czy użytkownik jest właścicielem notatki
+            if (notatki == null || notatki.UserId != User.Identity.Name)
             {
                 return NotFound();
             }
+
             return View(notatki);
         }
 
         // POST: Notatkis/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Tytul,Tresc,DataDodania,UserId")] Notatki notatki)
         {
-            if (id != notatki.Id)
+            if (id != notatki.Id) // Sprawdzenie id
             {
                 return NotFound();
             }
@@ -126,9 +152,10 @@ namespace KluczToSukcesDoKariery.Controllers
                 return NotFound();
             }
 
-            var notatki = await _context.Notatki
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (notatki == null)
+            var notatki = await _context.Notatki.FirstOrDefaultAsync(m => m.Id == id);
+
+            // Sprawdzenie, czy użytkownik jest właścicielem notatki
+            if (notatki == null || notatki.UserId != User.Identity.Name)
             {
                 return NotFound();
             }
@@ -141,23 +168,21 @@ namespace KluczToSukcesDoKariery.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Notatki == null)
-            {
-                return Problem("Entity set 'KluczToSukcesDoKarieryContext.Notatki'  is null.");
-            }
             var notatki = await _context.Notatki.FindAsync(id);
-            if (notatki != null)
+
+            // Sprawdzenie, czy użytkownik jest właścicielem notatki
+            if (notatki != null && notatki.UserId == User.Identity.Name)
             {
                 _context.Notatki.Remove(notatki);
+                await _context.SaveChangesAsync();
             }
-            
-            await _context.SaveChangesAsync();
+
             return RedirectToAction(nameof(Index));
         }
 
         private bool NotatkiExists(int id)
         {
-          return (_context.Notatki?.Any(e => e.Id == id)).GetValueOrDefault();
+            return (_context.Notatki?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
